+++ tests/integration/test_engine.py
@@ -51,21 +51,17 @@ class EngineSuite(unittest.TestCase):
             system_directory=os.getenv("SYSTEM_DIRECTORY", None),
         )
 
-        # Set
         i = Set(m, name="i", records=["seattle", "san-diego"])
         j = Set(m, name="j", records=["new-york", "chicago", "topeka"])
 
-        # Data
         a = Parameter(m, name="a", domain=[i], records=self.capacities)
         b = Parameter(m, name="b", domain=[j], records=self.demands)
         d = Parameter(m, name="d", domain=[i, j], records=self.distances)
         c = Parameter(m, name="c", domain=[i, j])
         c[i, j] = 90 * d[i, j] / 1000
 
-        # Variable
         x = Variable(m, name="x", domain=[i, j], type="Positive")
 
-        # Equation
         supply = Equation(m, name="supply", domain=[i])
         demand = Equation(m, name="demand", domain=[j])
 
@@ -128,21 +124,17 @@ class EngineSuite(unittest.TestCase):
             system_directory=os.getenv("SYSTEM_DIRECTORY", None),
         )
 
-        # Set
         i = Set(m, name="i", records=["seattle", "san-diego"])
         j = Set(m, name="j", records=["new-york", "chicago", "topeka"])
 
-        # Data
         a = Parameter(m, name="a", domain=[i], records=self.capacities)
         b = Parameter(m, name="b", domain=[j], records=self.demands)
         d = Parameter(m, name="d", domain=[i, j], records=self.distances)
         c = Parameter(m, name="c", domain=[i, j])
         c[i, j] = 90 * d[i, j] / 1000
 
-        # Variable
         x = Variable(m, name="x", domain=[i, j], type="Positive")
 
-        # Equation
         supply = Equation(m, name="supply", domain=[i])
         demand = Equation(m, name="demand", domain=[j])
 
@@ -166,21 +158,17 @@ class EngineSuite(unittest.TestCase):
             system_directory=os.getenv("SYSTEM_DIRECTORY", None),
         )
 
-        # Set
         i = Set(m, name="i", records=["seattle", "san-diego"])
         j = Set(m, name="j", records=["new-york", "chicago", "topeka"])
 
-        # Data
         a = Parameter(m, name="a", domain=[i], records=self.capacities)
         b = Parameter(m, name="b", domain=[j], records=self.demands)
         d = Parameter(m, name="d", domain=[i, j], records=self.distances)
         c = Parameter(m, name="c", domain=[i, j])
         c[i, j] = 90 * d[i, j] / 1000
 
-        # Variable
         x = Variable(m, name="x", domain=[i, j], type="Positive")
 
-        # Equation
         supply = Equation(m, name="supply", domain=[i])
         demand = Equation(m, name="demand", domain=[j])
 
@@ -229,21 +217,17 @@ class EngineSuite(unittest.TestCase):
             system_directory=os.getenv("SYSTEM_DIRECTORY", None),
         )
 
-        # Set
         i = Set(m, name="i", records=["seattle", "san-diego"])
         j = Set(m, name="j", records=["new-york", "chicago", "topeka"])
 
-        # Data
         a = Parameter(m, name="a", domain=[i], records=self.capacities)
         b = Parameter(m, name="b", domain=[j], records=self.demands)
         d = Parameter(m, name="d", domain=[i, j], records=self.distances)
         c = Parameter(m, name="c", domain=[i, j])
         c[i, j] = 90 * d[i, j] / 1000
 
-        # Variable
         x = Variable(m, name="x", domain=[i, j], type="Positive")
 
-        # Equation
         supply = Equation(m, name="supply", domain=[i])
         demand = Equation(m, name="demand", domain=[j])
 
@@ -274,21 +258,17 @@ class EngineSuite(unittest.TestCase):
             system_directory=os.getenv("SYSTEM_DIRECTORY", None),
         )
 
-        # Set
         i = Set(m, name="i", records=["seattle", "san-diego"])
         j = Set(m, name="j", records=["new-york", "chicago", "topeka"])
 
-        # Data
         a = Parameter(m, name="a", domain=[i], records=self.capacities)
         b = Parameter(m, name="b", domain=[j], records=self.demands)
         d = Parameter(m, name="d", domain=[i, j], records=self.distances)
         c = Parameter(m, name="c", domain=[i, j])
         c[i, j] = 90 * d[i, j] / 1000
 
-        # Variable
         x = Variable(m, name="x", domain=[i, j], type="Positive")
 
-        # Equation
         supply = Equation(m, name="supply", domain=[i])
         demand = Equation(m, name="demand", domain=[j])
 
@@ -309,24 +289,19 @@ class EngineSuite(unittest.TestCase):
     def test_non_blocking(self):
         m = Container(
             system_directory=os.getenv("SYSTEM_DIRECTORY", None),
-            working_directory=".",
         )
 
-        # Set
         i = Set(m, name="i", records=["seattle", "san-diego"])
         j = Set(m, name="j", records=["new-york", "chicago", "topeka"])
 
-        # Data
         a = Parameter(m, name="a", domain=[i], records=self.capacities)
         b = Parameter(m, name="b", domain=[j], records=self.demands)
         d = Parameter(m, name="d", domain=[i, j], records=self.distances)
         c = Parameter(m, name="c", domain=[i, j])
         c[i, j] = 90 * d[i, j] / 1000
 
-        # Variable
         x = Variable(m, name="x", domain=[i, j], type="Positive")
 
-        # Equation
         supply = Equation(m, name="supply", domain=[i])
         demand = Equation(m, name="demand", domain=[j])
 
@@ -348,20 +323,24 @@ class EngineSuite(unittest.TestCase):
             namespace=os.environ["ENGINE_NAMESPACE"],
             is_blocking=False,
         )
-        transport.solve(backend="engine", client=client)
+        # send jobs asynchronously
+        for _ in range(3):
+            transport.solve(backend="engine", client=client)
 
-        token = client.tokens[-1]
+        # gather the results
+        for i in range(3):
+            token = client.tokens[i]
 
-        job_status, _, exit_code = client.job.get(token)
-        while job_status != 10:
             job_status, _, exit_code = client.job.get(token)
+            while job_status != 10:
+                job_status, _, exit_code = client.job.get(token)
 
-        self.assertEqual(exit_code, 0)
+            self.assertEqual(exit_code, 0)
 
-        client.job.get_results(token, "out_dir")
+            client.job.get_results(token, f"tmp{os.sep}out_dir{i}")
 
         gdx_out_path = os.path.join(
-            "out_dir", os.path.basename(m.gdxOutputPath())
+            f"tmp{os.sep}out_dir0", os.path.basename(m.gdxOutputPath())
         )
         container = Container(load_from=gdx_out_path)
         self.assertTrue("x" in container.data)
@@ -377,11 +356,11 @@ class EngineSuite(unittest.TestCase):
             is_blocking=False,
         )
 
-        gms_path = os.path.join(os.getcwd(), "dummy.gms")
+        gms_path = os.path.join(os.getcwd(), "tmp", "dummy.gms")
         with open(gms_path, "w") as file:
             file.write("Set i / i1*i3 /;")
 
-        token = client.job.post(os.getcwd(), gms_path)
+        token = client.job.post(os.getcwd() + os.sep + "tmp", gms_path)
 
         status, _, _ = client.job.get(token)
         while status != 10:
@@ -419,11 +398,11 @@ class EngineSuite(unittest.TestCase):
             namespace=os.environ["ENGINE_NAMESPACE"],
             jwt=jwt_token,
         )
-        gms_path = os.path.join(os.getcwd(), "dummy2.gms")
+        gms_path = os.path.join(os.getcwd(), "tmp", "dummy2.gms")
         with open(gms_path, "w") as file:
             file.write("Set i / i1*i3 /;")
 
-        token = client.job.post(os.getcwd(), gms_path)
+        token = client.job.post(os.getcwd() + os.sep + "tmp", gms_path)
 
         status, _, _ = client.job.get(token)
         while status != 10: