+++ tests/integration/test_neos.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import os
 import unittest
 
+import pandas as pd
 from gamspy import (
     Container,
     Equation,
@@ -45,24 +46,19 @@ class NeosSuite(unittest.TestCase):
     def test_neos_blocking(self):
         m = Container(
             system_directory=os.getenv("SYSTEM_DIRECTORY", None),
-            working_directory=".",
         )
 
-        # Set
         i = Set(m, name="i", records=["seattle", "san-diego"])
         j = Set(m, name="j", records=["new-york", "chicago", "topeka"])
 
-        # Data
         a = Parameter(m, name="a", domain=[i], records=self.capacities)
         b = Parameter(m, name="b", domain=[j], records=self.demands)
         d = Parameter(m, name="d", domain=[i, j], records=self.distances)
         c = Parameter(m, name="c", domain=[i, j])
         c[i, j] = 90 * d[i, j] / 1000
 
-        # Variable
         x = Variable(m, name="x", domain=[i, j], type="Positive")
 
-        # Equation
         supply = Equation(m, name="supply", domain=[i])
         demand = Equation(m, name="demand", domain=[j])
 
@@ -80,7 +76,8 @@ class NeosSuite(unittest.TestCase):
         client = NeosClient(
             email=os.environ["NEOS_EMAIL"],
         )
-        transport.solve(backend="neos", client=client)
+        summary = transport.solve(backend="neos", client=client)
+        self.assertTrue(isinstance(summary, pd.DataFrame))
 
         import math
 
@@ -91,24 +88,19 @@ class NeosSuite(unittest.TestCase):
     def test_no_client(self):
         m = Container(
             system_directory=os.getenv("SYSTEM_DIRECTORY", None),
-            working_directory=".",
         )
 
-        # Set
         i = Set(m, name="i", records=self.canning_plants)
         j = Set(m, name="j", records=self.markets)
 
-        # Data
         a = Parameter(m, name="a", domain=[i], records=self.capacities)
         b = Parameter(m, name="b", domain=[j], records=self.demands)
         d = Parameter(m, name="d", domain=[i, j], records=self.distances)
         c = Parameter(m, name="c", domain=[i, j])
         c[i, j] = 90 * d[i, j] / 1000
 
-        # Variable
         x = Variable(m, name="x", domain=[i, j], type="Positive")
 
-        # Equation
         supply = Equation(m, name="supply", domain=[i])
         demand = Equation(m, name="demand", domain=[j])
 
@@ -131,21 +123,17 @@ class NeosSuite(unittest.TestCase):
             system_directory=os.getenv("SYSTEM_DIRECTORY", None),
         )
 
-        # Set
         i = Set(m, name="i", records=["seattle", "san-diego"])
         j = Set(m, name="j", records=["new-york", "chicago", "topeka"])
 
-        # Data
         a = Parameter(m, name="a", domain=[i], records=self.capacities)
         b = Parameter(m, name="b", domain=[j], records=self.demands)
         d = Parameter(m, name="d", domain=[i, j], records=self.distances)
         c = Parameter(m, name="c", domain=[i, j])
         c[i, j] = 90 * d[i, j] / 1000
 
-        # Variable
         x = Variable(m, name="x", domain=[i, j], type="Positive")
 
-        # Equation
         supply = Equation(m, name="supply", domain=[i])
         demand = Equation(m, name="demand", domain=[j])
 
@@ -174,24 +162,19 @@ class NeosSuite(unittest.TestCase):
     def test_neos_non_blocking(self):
         m = Container(
             system_directory=os.getenv("SYSTEM_DIRECTORY", None),
-            working_directory=".",
         )
 
-        # Set
         i = Set(m, name="i", records=["seattle", "san-diego"])
         j = Set(m, name="j", records=["new-york", "chicago", "topeka"])
 
-        # Data
         a = Parameter(m, name="a", domain=[i], records=self.capacities)
         b = Parameter(m, name="b", domain=[j], records=self.demands)
         d = Parameter(m, name="d", domain=[i, j], records=self.distances)
         c = Parameter(m, name="c", domain=[i, j])
         c[i, j] = 90 * d[i, j] / 1000
 
-        # Variable
         x = Variable(m, name="x", domain=[i, j], type="Positive")
 
-        # Equation
         supply = Equation(m, name="supply", domain=[i])
         demand = Equation(m, name="demand", domain=[j])
 
@@ -215,12 +198,14 @@ class NeosSuite(unittest.TestCase):
         job_number, job_password = client.jobs[-1]
         client.get_final_results(job_number, job_password)
         client.download_output(
-            job_number, job_password, working_directory="my_out_directory"
+            job_number,
+            job_password,
+            working_directory=f"tmp{os.sep}my_out_directory",
         )
 
         container = Container(
             system_directory=os.getenv("SYSTEM_DIRECTORY", None),
-            load_from="my_out_directory/output.gdx",
+            load_from=f"tmp{os.sep}my_out_directory/output.gdx",
         )
         self.assertTrue("x" in container.data)
         x.setRecords(container["x"].records)