+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 GAMSPy CHANGELOG
 ================
 
+GAMSPy 0.12.2
+-------------
+- General
+  - Add infeasibility_tolerance as a model attribute.
+  - Make urllib3 a true dependency instead of an optional one.
+  - Do not suppress compiler listing by default.
+  - Improve the performance of model attribute loading.
+  - Fix license path for model instance.
+- Documentation
+  - Add documentation about solver specific infeasibility options.
+
 GAMSPy 0.12.1
 -------------
 - General
@@ -32,7 +43,6 @@ GAMSPy 0.12.1
   - Remove documentation for delayed execution mode.
   - Add an example for providing solver options.
   - Document CLI for gamspy show commands.
-
 
 GAMSPy 0.12.0
 -------------